msm: camera: icp: Signal all fences when request is done.
Firmware notifies ICP KMD when all frames in payload are done,
so there is no need to keep track of buf dones individually.
Change made to signal all fences when firmware is done.

+++ drivers/media/platform/msm/camera/icp/icp_hw/icp_hw_mgr/cam_icp_hw_mgr.h
@@ -100,7 +100,7 @@ struct hfi_frame_process_info {
 	void *bitmap;
 	size_t bits;
 	struct mutex lock;
-	int32_t request_id[CAM_FRAME_CMD_MAX];
+	uint64_t request_id[CAM_FRAME_CMD_MAX];
 	uint32_t num_out_resources[CAM_FRAME_CMD_MAX];
 	uint32_t out_resource[CAM_FRAME_CMD_MAX][CAM_MAX_OUT_RES];
 };
